fix: handle spaces in string inputs
Using Scanln causes string values with spaces to only be partially read
because just a single variable is passed to store the line scan result.

This commit fixes the problem by reading stdin into a reader buffer and
then reading a full line with the input item value is a string pointer.